refactor: moved DevicesRepo methods to IdentitiesRepo and deleted the first
