+++ Modules/Devices/src/Devices.Application/Infrastructure/Persistence/Repository/IIdentitiesRepository.cs
@@ -7,9 +7,20 @@ namespace Backbone.Modules.Devices.Application.Infrastructure.Persistence.Reposi
 
 public interface IIdentitiesRepository
 {
+    #region Identities
     Task<DbPaginationResult<Identity>> FindAll(PaginationFilter paginationFilter);
 #nullable enable
     Task<Identity?> FindByAddress(IdentityAddress address, CancellationToken cancellationToken);
 #nullable disable
+    #endregion
+
+    #region Users
     Task AddUserForIdentity(ApplicationUser user, string password);
+    #endregion
+
+    #region Devices
+    Task<DbPaginationResult<Device>> FindAll(IdentityAddress identity, IEnumerable<DeviceId> ids, PaginationFilter paginationFilter);
+    Task<Device> GetDeviceById(DeviceId deviceId, CancellationToken cancellationToken, bool track = false);
+    Task Update(Device device, CancellationToken cancellationToken);
+    #endregion
 }
