+++ Modules/Devices/src/Devices.Infrastructure/Persistence/IServiceCollectionExtensions.cs
@@ -56,7 +56,6 @@ public static class IServiceCollectionExtensions
     {
         services.AddTransient<IIdentitiesRepository, IdentitiesRepository>();
         services.AddTransient<ITiersRepository, TiersRepository>();
-        services.AddTransient<IDevicesRepository, DevicesRepository>();
         services.AddTransient<IChallengesRepository, ChallengesRepository>();
     }
